microservices/doc-ingestor/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
psycopg2-binary==2.9.11
pika==1.3.2
python-dotenv==1.0.0
tika==2.6.0
pytesseract==0.3.10
Pillow
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0
pydantic
pydantic-settings
aiofiles==23.2.1
